change main.rb to app.rb
change main.rb to app.rb and create main.rb just as initialization file
